Counting timed out jobs as failed.
Results that exist in the results directory will be considered failed if
the last time they started (or the last time they were touched for older
experiment sets) is further in the past than the walltime in the
expfile. This allows for the `summary` and `watch` commands to
accurately represented timed out jobs as failed.